EVERYTHING
everything works except mission 3 and 4 don't have their pass-key's
hidden in the pages yet. Accounts are now successfully created, need to
be logged into (or may break the site), and all hint and pass-key
functions work. mission 5 now works and is renamed "etisoppo.php" and
the idea is that you change the url to "opposite.php" which gives you
the pass-key. after a mission is completed, there is a slight fluke
where you need to refresh the page or go to a new page before the next
mission tab becomes clickable.